Calculation methods

+++ migrations/2018-09-21-175400_create_sale/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE sales(
     id SERIAL PRIMARY KEY,
     client_id INTEGER NOT NULL REFERENCES clients(id),
     sale_date DATE NOT NULL,
-    sub_total INTEGER,
-    total INTEGER,
+    sub_total FLOAT,
+    total FLOAT,
     observation TEXT
 );
